requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
peft
albumentations
opencv-python
pillow
numpy
pandas
matplotlib
seaborn
tqdm
scikit-learn
torcheval